clean: delete old python project
