+++ Cargo.toml
@@ -3,6 +3,10 @@ name = "dragon-runner-rs"
 version = "0.1.0"
 edition = "2021"
 
+[[bin]]
+name = "dragon-runner"
+path = "src/main.rs"
+
 [dependencies]
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
