--- pyproject.toml
@@ -1,26 +1,0 @@
-[build-system]
-requires = ["setuptools>=45", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "dragon-runner"
-version = "1.0.0"
-description = "An experimental successor to the 415 tester"
-authors = [{name = "Justin Meimar", email = "meimar@ualberta.ca"}]
-license = {file = "LICENSE"}
-readme = "README.md"
-requires-python = ">=3.6"
-classifiers = ["Tester"]
-dependencies = [
-    "colorama==0.4.6",
-    "pytest==8.3.3",
-    "numpy==2.2.4",
-    "Flask==3.1.0",
-    'flask-cors==6.0.0'
-]
-
-[project.scripts]
-dragon-runner = "dragon_runner.src.main:main"
-
-[tool.setuptools.package-data]
-dragon_runner = ["py.typed"]
