+++ scripts/gen_config.py
