+++ src/config.rs
@@ -392,3 +392,95 @@ pub fn load_config(config_path: &str, args: Option<&RunnerArgs>) -> Option<Confi
 
     Some(Config::new(config_path, &config_data, debug_package, package_filter))
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn configs_dir() -> PathBuf {
+        PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("tests").join("configs")
+    }
+
+    fn config_path(name: &str) -> String {
+        configs_dir().join(name).to_string_lossy().into_owned()
+    }
+
+    #[test]
+    fn test_valid_config() {
+        let path = config_path("gccPassConfig.json");
+        let config = load_config(&path, None).expect("config should load");
+
+        assert!(
+            Path::new(&config.test_dir).exists(),
+            "test_dir should exist: {}",
+            config.test_dir
+        );
+        assert!(!config.packages.is_empty(), "should have packages");
+
+        for pkg in &config.packages {
+            assert!(!pkg.subpackages.is_empty(), "package {} should have subpackages", pkg.name);
+            for spkg in &pkg.subpackages {
+                assert!(!spkg.tests.is_empty(), "subpackage {} should have tests", spkg.name);
+            }
+        }
+
+        assert!(config.errors.is_empty(), "should have no errors");
+    }
+
+    #[test]
+    fn test_package_filter() {
+        let path = config_path("gccPassConfig.json");
+        let config = load_config(&path, None).expect("config should load");
+
+        let all_subpackages: Vec<&str> = config
+            .packages
+            .iter()
+            .flat_map(|pkg| pkg.subpackages.iter())
+            .map(|spkg| spkg.path.as_str())
+            .collect();
+
+        assert!(!all_subpackages.is_empty(), "should have subpackages");
+
+        let filter_pattern = "*ErrorPass*";
+        let filtered: Vec<&&str> = all_subpackages
+            .iter()
+            .filter(|path| {
+                glob::Pattern::new(&filter_pattern.to_lowercase())
+                    .map(|pat| pat.matches(&path.to_lowercase()))
+                    .unwrap_or(false)
+            })
+            .collect();
+
+        assert!(!filtered.is_empty(), "filter should match some subpackages");
+
+        for path in &filtered {
+            assert!(
+                path.to_lowercase().contains("errorpass"),
+                "filtered path should contain 'errorpass': {}",
+                path
+            );
+        }
+    }
+
+    #[test]
+    fn test_invalid_dir_config() {
+        let path = config_path("invalidDirConfig.json");
+        let config = load_config(&path, None).expect("config should load");
+
+        assert!(!config.errors.is_empty(), "should have errors for invalid dir");
+        assert!(!Path::new(&config.test_dir).exists(), "test_dir should not exist");
+    }
+
+    #[test]
+    fn test_invalid_exe_config() {
+        let path = config_path("invalidExeConfig.json");
+        let config = load_config(&path, None).expect("config should load");
+
+        assert!(!config.errors.is_empty(), "should have errors for invalid exe");
+        assert_eq!(config.executables.len(), 1);
+        assert!(
+            !Path::new(&config.executables[0].exe_path).exists(),
+            "exe_path should not exist"
+        );
+    }
+}
